agregadas alertas de envío de email con link, de email existente y de envío de email de registro exitoso
