+++ api/src/controllers/auth.controller.js
@@ -28,10 +28,9 @@ const validate = async (req, res) => {
       res.status(500).send(error.message)
     } else {
       console.log('email enviado')
-      res.status(200).jsonp(token)
+      return res.status(200).send()
     }
   })
-    
 
   } catch (error) {
     return res.status(400).json({ error: error.message });
